requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dateutil
